+++ packages/primitive-contracts/contracts/option/interfaces/IFactory.sol
@@ -1,8 +1,15 @@
 pragma solidity ^0.6.2;
 
 interface IFactory {
-    function deploy(address tokenU, address tokenS, uint base, uint quote, uint expiry)
-        external returns (address);
+    function deploy(
+        address tokenU,
+        address tokenS,
+        uint256 base,
+        uint256 quote,
+        uint256 expiry
+    ) external returns (address);
+
     function kill(address option) external;
+
     function initialize(address option, address redeem) external;
 }
