+++ packages/primitive-contracts/contracts/option/interfaces/IFlash.sol
@@ -1,5 +1,9 @@
 pragma solidity ^0.6.2;
 
 interface IFlash {
-    function primitiveFlash(address receiver, uint outTokenU, bytes calldata data) external; 
+    function primitiveFlash(
+        address receiver,
+        uint256 outTokenU,
+        bytes calldata data
+    ) external;
 }
