+++ packages/primitive-contracts/contracts/pool/applications/AMM.sol
@@ -14,22 +14,22 @@ import "../interfaces/IUniswapV2Factory.sol";
 import "../interfaces/IUniswapV2Router01.sol";
 
 contract AMM is Pool {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
-    uint public constant MANTISSA = 10**36;
-    uint public constant SLIPPAGE = 10**10;
-    uint public constant ONE_ETHER = 1 ether;
-    uint public constant DISCOUNT_RATE = 5;
-    uint public constant MIN_VOLATILITY = 10**15;
-    uint public volatility;
+    uint256 public constant MANTISSA = 10**36;
+    uint256 public constant SLIPPAGE = 10**10;
+    uint256 public constant ONE_ETHER = 1 ether;
+    uint256 public constant DISCOUNT_RATE = 5;
+    uint256 public constant MIN_VOLATILITY = 10**15;
+    uint256 public volatility;
 
     address public oracle;
     address public weth;
     address public router;
 
     event Market(address tokenP);
-    event Buy(address indexed from, uint outTokenU, uint premium);
-    event Sell(address indexed from, uint inTokenP, uint premium);
+    event Buy(address indexed from, uint256 outTokenU, uint256 premium);
+    event Sell(address indexed from, uint256 inTokenP, uint256 premium);
 
     constructor(
         address _weth,
@@ -37,10 +37,7 @@ contract AMM is Pool {
         address _oracle,
         address _factory,
         address _router
-    )
-        public
-        Pool(_tokenP, _factory)
-    {
+    ) public Pool(_tokenP, _factory) {
         weth = _weth;
         oracle = _oracle;
         router = _router;
@@ -52,12 +49,15 @@ contract AMM is Pool {
      * @dev Accepts deposits of underlying tokens.
      * @param inTokenU Quantity of underlyings to deposit.
      */
-    function deposit(uint inTokenU) external whenNotPaused nonReentrant
-        returns (uint outTokenPULP, bool success)
+    function deposit(uint256 inTokenU)
+        external
+        whenNotPaused
+        nonReentrant
+        returns (uint256 outTokenPULP, bool success)
     {
         address _tokenP = tokenP;
         address tokenU = IOption(_tokenP).tokenU();
-        (uint totalBalance) = totalPoolBalance(_tokenP);
+        uint256 totalBalance = totalPoolBalance(_tokenP);
         (outTokenPULP) = _addLiquidity(msg.sender, inTokenU, totalBalance);
         require(
             IERC20(tokenU).transferFrom(msg.sender, address(this), inTokenU),
@@ -70,18 +70,30 @@ contract AMM is Pool {
      * @dev Withdraws underlyings proportional to liquidity tokens burned.
      * @param inTokenPULP Quantity of liquidity tokens to burn.
      */
-    function withdraw(uint inTokenPULP) external whenNotPaused nonReentrant returns (bool) {
+    function withdraw(uint256 inTokenPULP)
+        external
+        whenNotPaused
+        nonReentrant
+        returns (bool)
+    {
         // Store for gas savings.
         address _tokenP = tokenP;
-        (address tokenU, address tokenS, address tokenR, , ,) = IOption(_tokenP).getOption();
-        (uint totalBalance) = totalPoolBalance(_tokenP);
-        (uint balanceU,) = balances();
+        (address tokenU, address tokenS, address tokenR, , , ) = IOption(
+            _tokenP
+        )
+            .getOption();
+        uint256 totalBalance = totalPoolBalance(_tokenP);
+        (uint256 balanceU, ) = balances();
 
         // Burn liquidity tokens.
-        (uint outTokenU) = _removeLiquidity(msg.sender, inTokenPULP, totalBalance);
+        uint256 outTokenU = _removeLiquidity(
+            msg.sender,
+            inTokenPULP,
+            totalBalance
+        );
 
         // If not enough available liquidity to draw, redeem and swap strike tokens.
-        if(balanceU < outTokenU) {
+        if (balanceU < outTokenU) {
             _redeemAndSwapStrike(_tokenP, tokenU, tokenS, tokenR);
         }
         require(balanceU >= outTokenU, "ERR_BAL_INSUFFICIENT");
@@ -94,21 +106,25 @@ contract AMM is Pool {
      * Will revert if maxDraw is 0.
      * @param _tokenP Address of  option contract.
      */
-    function _redeemAndSwapStrike(address _tokenP, address tokenU, address tokenS, address tokenR)
-        private
-        returns (uint outTokenR)
-    {
+    function _redeemAndSwapStrike(
+        address _tokenP,
+        address tokenU,
+        address tokenS,
+        address tokenR
+    ) private returns (uint256 outTokenR) {
         // Check how many tokenS can be pulled from Option.sol.
-        uint balanceR = IERC20(tokenR).balanceOf(address(this));
-        uint cacheS = IOption(_tokenP).cacheS();
-        uint maxDraw = balanceR > cacheS ? cacheS : balanceR;
+        uint256 balanceR = IERC20(tokenR).balanceOf(address(this));
+        uint256 cacheS = IOption(_tokenP).cacheS();
+        uint256 maxDraw = balanceR > cacheS ? cacheS : balanceR;
 
         // Redeem tokens.
         (outTokenR) = _redeem(address(this), maxDraw);
         assert(outTokenR == maxDraw);
 
-        uint market = IOracle(oracle).marketPrice();
-        uint minOut = tokenS == weth ? market : outTokenR.mul(ONE_ETHER).div(market);
+        uint256 market = IOracle(oracle).marketPrice();
+        uint256 minOut = tokenS == weth
+            ? market
+            : outTokenR.mul(ONE_ETHER).div(market);
 
         address[] memory path = new address[](2);
         path[0] = tokenS;
@@ -128,16 +144,16 @@ contract AMM is Pool {
      * @param outTokenP The quantity of options to buy, which allow the purchase of 1:1 tokenU.
      * @return bool True if the msg.sender receives tokenP.
      */
-    function buy(uint outTokenP) external nonReentrant returns (bool) {
+    function buy(uint256 outTokenP) external nonReentrant returns (bool) {
         // Store in memory for gas savings.
         address _tokenP = tokenP;
         (
             address tokenU, // Assume DAI.
-            address tokenS, // Assume ETH.
-             , // Assume tokenR and we don't need it in this function.
-            uint base,
-            uint quote,
-            uint expiry
+            address tokenS, // Assume ETH. // Assume tokenR and we don't need it in this function.
+            ,
+            uint256 base,
+            uint256 quote,
+            uint256 expiry
         ) = IOption(_tokenP).getOption();
 
         // Optimistically mint option tokens to the msg.sender.
@@ -145,9 +161,9 @@ contract AMM is Pool {
 
         // Calculates and then updates the volatility global state variable.
         volatility = calculateVolatilityProxy(_tokenP);
-        
+
         // Calculate premium. Denominated in tokenU PER tokenS 'covered'.
-        (uint premium) = IOracle(oracle).calculatePremium(
+        uint256 premium = IOracle(oracle).calculatePremium(
             tokenU,
             tokenS,
             volatility,
@@ -155,10 +171,10 @@ contract AMM is Pool {
             quote,
             expiry
         );
-        
+
         // Calculate total premium to pay. Premium should be in underlying token units.
         premium = outTokenP.mul(premium).div(ONE_ETHER);
-        if(tokenU == weth) premium = MANTISSA.div(premium);
+        if (tokenU == weth) premium = MANTISSA.div(premium);
         require(premium > 0, "ERR_PREMIUM_ZERO");
 
         // Pulls payment in tokenU from msg.sender and then pushes tokenP (option).
@@ -173,27 +189,26 @@ contract AMM is Pool {
      * @notice The pool buys options at a discounted rate based on the current premium quote.
      * @param inTokenP The amount of  option tokens that are being sold.
      */
-    function sell(uint inTokenP) external nonReentrant returns (bool) {
+    function sell(uint256 inTokenP) external nonReentrant returns (bool) {
         // Store in memory for gas savings.
         address _tokenP = tokenP;
         (
             address tokenU, // Assume DAI.
             address tokenS, // Assume ETH.
             address tokenR, // Assume tokenR.
-            uint base,
-            uint quote,
-            uint expiry
+            uint256 base,
+            uint256 quote,
+            uint256 expiry
         ) = IOption(_tokenP).getOption();
 
         // Check tokenP balance.
         require(
-            IERC20(_tokenP).balanceOf(msg.sender) >= inTokenP &&
-            inTokenP > 0,
+            IERC20(_tokenP).balanceOf(msg.sender) >= inTokenP && inTokenP > 0,
             "ERR_BAL_PRIME"
         );
 
         // Calculate the current premium quote.
-        (uint premium) = IOracle(oracle).calculatePremium(
+        uint256 premium = IOracle(oracle).calculatePremium(
             tokenU,
             tokenS,
             volatility,
@@ -208,17 +223,25 @@ contract AMM is Pool {
         // Calculate total premium.
         // Units: tokenU * (tokenU / tokenS) / 10^18 units = total quantity tokenU quote.
         premium = inTokenP.mul(premium).div(ONE_ETHER);
-        if(tokenU == weth) { premium = MANTISSA.div(premium); }
+        if (tokenU == weth) {
+            premium = MANTISSA.div(premium);
+        }
 
         // Check to see if pool has the premium to pay out.
-        require(IERC20(tokenU).balanceOf(address(this)) >= premium, "ERR_BAL_UNDERLYING");
-        
+        require(
+            IERC20(tokenU).balanceOf(address(this)) >= premium,
+            "ERR_BAL_UNDERLYING"
+        );
+
         // Calculate amount of redeem needed to close position with inTokenU.
-        uint outTokenR = inTokenP.mul(quote).div(base);
-        require(IERC20(tokenR).balanceOf(address(this)) >= outTokenR, "ERR_BAL_REDEEM");
+        uint256 outTokenR = inTokenP.mul(quote).div(base);
+        require(
+            IERC20(tokenR).balanceOf(address(this)) >= outTokenR,
+            "ERR_BAL_REDEEM"
+        );
 
         // Call the close function to close the option position and receive underlyings.
-        (uint outTokenU) = _close(outTokenR, inTokenP);
+        uint256 outTokenU = _close(outTokenR, inTokenP);
         assert(inTokenP >= outTokenU);
 
         // Pay out the total premium to the seller.
@@ -230,20 +253,28 @@ contract AMM is Pool {
      * @dev Calculates the Pool's Utilization to use as a proxy for volatility.
      * @notice If Pool is not utilized at all, the default volatility is 250.
      */
-    function calculateVolatilityProxy(address _tokenP) public view returns (uint _volatility) {
-        (uint utilized) = totalUtilized(_tokenP);
-        uint totalBalance = totalPoolBalance(_tokenP);
-        if(totalBalance > 0) _volatility = utilized.mul(ONE_ETHER).div(totalBalance); // Volatility with 1e18 decimals.
-        if(_volatility < MIN_VOLATILITY) {
+    function calculateVolatilityProxy(address _tokenP)
+        public
+        view
+        returns (uint256 _volatility)
+    {
+        uint256 utilized = totalUtilized(_tokenP);
+        uint256 totalBalance = totalPoolBalance(_tokenP);
+        if (totalBalance > 0)
+            _volatility = utilized.mul(ONE_ETHER).div(totalBalance); // Volatility with 1e18 decimals.
+        if (_volatility < MIN_VOLATILITY) {
             _volatility = 1000;
         } else _volatility = _volatility.div(MIN_VOLATILITY).add(1000);
-
     }
 
     /**
      * @dev Calculates the amount of utilized tokenU assets outstanding.
      */
-    function totalUtilized(address _tokenP) public view returns (uint utilized) {
+    function totalUtilized(address _tokenP)
+        public
+        view
+        returns (uint256 utilized)
+    {
         // Assume tokenR is proportional to tokenS (weth) at a 1:1 ratio.
         // TokenR is always minted proportionally to the ratio between tokenU and tokenS (strike quote).
         // Assume a ratio of 200 DAI per 1 ETH.
@@ -253,14 +284,19 @@ contract AMM is Pool {
         // TokenR is redeemable to tokenS at a 1:1 ratio (1 tokenR can be redeemed for 1 weth).
         // The utilized amount of tokenU is therefore this calculation:
         // (tokenR = tokenS = weth) * Quantity of tokenU (base) / Quantity of tokenS (quote).
-        ( , , address tokenR, uint base, uint quote, ) = IOption(_tokenP).getOption();
+        (, , address tokenR, uint256 base, uint256 quote, ) = IOption(_tokenP)
+            .getOption();
         utilized = IERC20(tokenR).balanceOf(address(this)).mul(base).div(quote);
     }
 
     /**
      * @dev Returns the contract balance of tokenU.
      */
-    function totalUnutilized(address _tokenP) public view returns (uint balanceU) {
+    function totalUnutilized(address _tokenP)
+        public
+        view
+        returns (uint256 balanceU)
+    {
         // The unutilized balance is equal to the balance of tokenU held by the pool.
         balanceU = IERC20(IOption(_tokenP).tokenU()).balanceOf(address(this));
     }
@@ -269,11 +305,13 @@ contract AMM is Pool {
      * @dev Returns to total balance of tokenU that the contract has accross accounts.
      * @notice Total Balance = utilized + unutilized. Utilized = underwritten tokenU.
      */
-    function totalPoolBalance(address _tokenP) public view returns (uint totalBalance) {
+    function totalPoolBalance(address _tokenP)
+        public
+        view
+        returns (uint256 totalBalance)
+    {
         // Unutilized is the balance of tokenU in the contract. Utilized is outstanding tokenU.
         // Utilized assets are held in the  contract waiting to be exercised or expired.
         totalBalance = totalUnutilized(_tokenP).add(totalUtilized(_tokenP));
     }
 }
-
-
